Refactor card models into dedicated modules
